requirements.txt
configparser==7.1.0
controller==0.1.0
gym==0.25.1
gymnasium==0.29.1
lxml==5.3.0
wandb==0.17.6
matplotlib==3.9.2
memory_profiler==0.61.0
numpy==1.25.0
opencv_python==4.9.0.80
openpyxl==3.1.5
pandas==2.2.2
Pillow==10.4.0
psutil==6.0.0
pyproj==3.6.1
pyserial==3.5
sb3_contrib==2.3.0
setproctitle==1.3.3
shapely==2.0.6
stable_baselines3==2.3.0
sumolib==1.20.0
tensorboard==2.17.1
torch==2.2.0
torchaudio==2.2.0
torchsummary==1.5.1
torchvision==0.17.0
tqdm==4.66.5
traci==1.20.0
transforms3d==0.4.2
wandb==0.17.6
webcolors==24.8.0
